requirements.txt
amqp==5.1.1
aniso8601==7.0.0
anyio==3.6.1
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.11.3
asttokens==2.2.1
attrs==23.1.0
autopep8==1.4.4

crispy-bootstrap5==0.6
cryptography==39.0.0
cssselect2==0.6.0
dateparser==1.1.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.6.0
dill==0.3.4
distlib==0.3.4
dj-database-url==1.2.0
Django==4.0.5

django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-flower==1.0.0
django-grappelli==3.0.4
django-heroku==0.3.1
django-hosts==5.2
django-htmx==1.8.0
django-jazzmin==2.5.0
django-jinja==2.10.2
django-js-asset==2.0.0
django-like-systems==0.0.1
django-localflavor==3.1
django-on-heroku==1.1.2
django-pagination-bootstrap==2.4.6
django-parler==2.3
django-paypal==2.0
django-rosetta==0.9.8
django-social-share==2.3.0
django-social-widgets==0.5.0
django-taggit==3.0.0
django-timezone-field==4.2.3
django-tinymce==3.6.1
django-utils-six==2.0
Django-Verify-Email==2.0.3
django-weasyprint==2.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.0
djorm-ext-pgtrgm==0.3

graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0

humanize==0.5.1



Pillow==9.1.1
pipenv==2022.6.7


psycopg2-binary==2.9.7




tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
twilio==7.11.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uri-template==1.2.0
uritools==4.0.0
urllib3==1.26.15
validate-email==1.3
vine==5.0.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
weasyprint==56.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.3.2
wget==3.2
whitenoise==6.2.0
widgetsnbextension==4.0.7
wikipedia==1.4.0
wrapt==1.11.2
xhtml2pdf==0.2.8
youtube-search-python==1.6.6
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.2.1
